+++ ImageFeed/Auth/AuthViewController.swift
@@ -17,9 +17,7 @@ final class AuthViewController: UIViewController {
                 let webViewViewController = segue.destination
                     as? WebViewViewController
             else {
-                assertionFailure(
-                    "Failed to prepare for \(showWebViewSegueIdentifier)"
-                )
+                assertionFailure("Failed to prepare for \(showWebViewSegueIdentifier)")
                 return
             }
             webViewViewController.delegate = self
@@ -29,20 +27,10 @@ final class AuthViewController: UIViewController {
     }
 
     private func configureBackButton() {
-        navigationController?.navigationBar.backIndicatorImage = UIImage(
-            named: "nav_back_button"
-        )
-        navigationController?.navigationBar.backIndicatorTransitionMaskImage =
-            UIImage(named: "nav_back_button")
-        navigationItem.backBarButtonItem = UIBarButtonItem(
-            title: "",
-            style: .plain,
-            target: nil,
-            action: nil
-        )
-        navigationItem.backBarButtonItem?.tintColor = UIColor(
-            named: "YP Black (iOS)"
-        )
+        navigationController?.navigationBar.backIndicatorImage = UIImage(named: "nav_back_button")
+        navigationController?.navigationBar.backIndicatorTransitionMaskImage = UIImage(named: "nav_back_button")
+        navigationItem.backBarButtonItem = UIBarButtonItem(title: "",style: .plain,target: nil,action: nil)
+        navigationItem.backBarButtonItem?.tintColor = UIColor(named: "YP Black (iOS)")
     }
 }
 extension AuthViewController: WebViewViewControllerDelegate {
