+++ ImageFeed/Auth/AuthViewControllerDelegate.swift
@@ -3,5 +3,4 @@ import UIKit
 protocol AuthViewControllerDelegate: AnyObject {
 
     func didAuthenticate(_ vc: AuthViewController)
-
 }
