+++ ImageFeed/Auth/NetworkError.swift
@@ -1,0 +1,9 @@
+import UIKit
+
+enum NetworkError: Error {
+    case httpStatusCode(Int)
+    case urlRequestError(Error)
+    case urlSessionError
+    case invalidRequest
+    case decodingError(Error)
+}
