+++ ImageFeed/Auth/OAuth2Service.swift
@@ -5,10 +5,7 @@ final class OAuth2Service {
     private init() {}
     private let tokenStorage = OAuth2TokenStorage()
 
-    func fetchOAuthToken(
-        code: String,
-        completion: @escaping (Result<String, Error>) -> Void
-    ) {
+    func fetchOAuthToken(code: String,completion: @escaping (Result<String, Error>) -> Void) {
         guard let request = makeOAuthTokenRequest(code: code) else {
             print("Запрос не создан")
             return
@@ -18,12 +15,8 @@ final class OAuth2Service {
             case .success(let data):
                 do {
                     let decoder = JSONDecoder()
-                    let responseBody = try decoder.decode(
-                        OAuthTokenResponseBody.self,
-                        from: data
-                    )
+                    let responseBody = try decoder.decode(OAuthTokenResponseBody.self,from: data)
                     self.tokenStorage.token = responseBody.access_token
-
                     completion(.success(responseBody.access_token))
                 } catch {
                     print("не получилось расшифровать данные")
@@ -38,10 +31,7 @@ final class OAuth2Service {
     }
 
     private func makeOAuthTokenRequest(code: String) -> URLRequest? {
-        guard
-            var urlComponents = URLComponents(
-                string: "https://unsplash.com/oauth/token"
-            )
+        guard var urlComponents = URLComponents(string: "https://unsplash.com/oauth/token")
         else {
             return nil
         }
