+++ ImageFeed/Auth/URLSession+data.swift.swift
@@ -1,18 +1,7 @@
 import UIKit
 
-enum NetworkError: Error {
-    case httpStatusCode(Int)
-    case urlRequestError(Error)
-    case urlSessionError
-    case invalidRequest
-    case decodingError(Error)
-}
-
 extension URLSession {
-    func data(
-        for request: URLRequest,
-        completion: @escaping (Result<Data, Error>) -> Void
-    ) -> URLSessionTask {
+    func data(for request: URLRequest,completion: @escaping (Result<Data, Error>) -> Void) -> URLSessionTask {
         let fulfillCompletionOnTheMainThread: (Result<Data, Error>) -> Void = {
             result in
             DispatchQueue.main.async {
@@ -44,7 +33,6 @@ extension URLSession {
                 }
             }
         )
-
         return task
     }
 }
