+++ ImageFeed/SingleImage/SingleImageViewController.swift
@@ -63,6 +63,6 @@ final class SingleImageViewController: UIViewController {
 
 extension SingleImageViewController: UIScrollViewDelegate {
     func viewForZooming(in scrollView: UIScrollView) -> UIView? {
-        return imageView
+         imageView
     }
 }
